+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 
 * Remove redundant API value from big number component ([PR #2459](https://github.com/alphagov/govuk_publishing_components/pull/2459))
 * Remove use of `govuk-list` from navigation header ([PR #2460](https://github.com/alphagov/govuk_publishing_components/pull/2460))
+* Extend new scroll tracker functionality ([PR #2411](https://github.com/alphagov/govuk_publishing_components/pull/2411))
 
 ## 27.14.0